Add `--save` to `kubernetes create`
Allows a user to immediately save their configuration of a newly-created
cluster assuming they use --wait as a switch - cleanly warns the user that
the config could not be saved if they do not use --wait.

Also checks for presence of `~/.kube/` directory to prevent erroring in
file write of config if it doesn't.