Use only cluster by default (#40)
* Implement check to see whether user has multiple clusters

If a user does not specify a cluster to install a kubernetes app to, checks
if the user has multiple clusters. If only one cluster is found, installs
specified app to that cluster. If multiple clusters found, lets user know.
Closes #38 when merged

* Simplify logic if no cluster name provided in `app add`

Logic for determining cluster number (and returning the only one as default
if count == 1 is now in Finder.detect_cluster and passed argument is blank